fix NoMethodError for serialized attributes
serialized_attributes is no longer available on AR instances and is
deprecated on AR classes. stole code from
https://github.com/thoughtbot/shoulda-matchers/commit/547d8b762f6c880d09eec3781a3a53770578175b
to work around this.

Communication around the deprecation suggests that attributes with
a serializer should behave like any other so it feels like there should
be a better solution but this at least works around the problem