deprecate the map option for load_db
ActiveRecord::FixtureSet used to let you specify the target table
name on initialization. It now uses the table name from the class
if it can find one.

Since our use case is extracting data from AR objects and
reloading them, ActiveRecord will always find the class's table name
and ignore the map option. This is ruby so we can force the FixtureSet
to use the new name, but it falls outside the use case of this gem. This is
not an ETL library, it is a library to generate test data.