Add fact method for SDP usage
The version of this gem that SDP uses is apparently custom. I can't
find it in either the public GitHub repo or the private enterprise one.
The one difference I see in SDPs version is that it supports the `fact`
method. You can write a test like

```
expect(x).to eq(Fe.fact(:fixture_name, :fact_attribute_name))
```

and in your fixture's manifest file you have

```
:fact_hash:
  :fact_attribute_name: 200
```

Then, when you run the test you can test that x is equal to 200

There is probably also code for extracting those facts, but I have not
ported that because I don't currently see a need. All I need is the
ability to pull those facts back out in my tests.