Cleanup dev gems, move to using gemspec
Current standard is to put runtime and development gems in the gemspec,
so I'm doing that here. I've removed the debugger gem because I find it
to be more trouble that its worth.

I'm locking the ruby version to 1.9.3 for now. But that will change once
I get travis setup to test the gem under multiple Ruby versions.