Add `fact_hash` method to extractor
I'm giving extractor the ability to return a fact_hash because that is
what was done in the now-unused version of this code that implemented
fact_hash. It may not be the best place for this code to live long term,
but it's fine for now.

I've added the barest-of-bones test for this code in a new spec file.
The existing specs are more along what I'd call 'feature' specs, but I
didn't see anything that follows the common Rspec model of having unit
tests in a directory that mirrors your /lib directory. So I'm starting
that now with a test for /lib/fe/extractor.rb in
/spec/lib/fe/extractor_spec.rb

As time goes on we can migrate/create unit tests there.