Add `build` method for Extractor
A very common pattern I see in the gem is:

```
extractor = Fe::Extractor.new
extractor.name = 'extract_name'
=>....a bunch of stuff done with the extractor
```

I'm wrapping this up in a `build` method that accepts an extract name
and returns the extractor instance with the name set.

Other than saving some lines of code in other methods, this is also
easier to test. Now I can just stub the `build` method instead of having
to stub both the initializer and the name setter.