Create top-level methods for setting/getting facts
This is a Iron Fixture Extractor method that is used by SDP and was part
of the 1.2.0 code used in that project.

While I'm not sure that the Fe module is the best place for these
methods to live, I'm going to go along with it for now.

These methods act as simple wrappers around Extractor methods, so the
tests are only checking that the methods we want to call are called.