Allow building of extractor with dependency injection
Most client code that builds an Extractor now does so like this:

```
extractor = Fe::Extractor.build(name)
extractor.load_from_manifest
```

This pattern is both common and hard to test, since the
`load_from_manifest` method opens a file using methods internal to
Extractor.

Since this pattern is so common, I'm creating a declarative_builder for
it, `build_from_manifest` which accepts a name and an optional manifest
file.

For existing client code, it can pass in no manifest file and the YAML
file is loaded as it always has been.

For testing code, however, we can now inject our own manifest hashes.
This makes testing much easier.

Since clients no longer deal with `load_from_manifest` directly, I can
remove the guarding clause in that method.

There's a possibly weird behavior in `load_from_manifest` that I have
kept, since I don't want to introduce breaking changes in this release.
If client code sets the extractor's name to a value, that name can be
overwritten by what is in the manifest hash. I do not see how this could
happen, honestly. And if it did happen that would probably be bad, as
the 'name' is pretty vital to how Extractor works. However, the code
exists there now and I'm not going to take it out until there's a major
release.