Use attributes api to get serialized attributes
ActiveRecord Column objects no longer respond to `cast_type`. Use the
attributes API methods to do our reflection instead.

I want to say we should be able to get rid of this reflection entirely
and use the `ActiveRecord::Attribute#value_for_database` and trust it
will do the right thing, but I have not yet found the right way to
access that method.